refactor(examples): 迁移 keyboard 至 plugins 目录
