+++ examples/src/App.tsx
@@ -34,7 +34,7 @@ import { CaseSwimlaneExample } from './pages/case/swimlane'
 import { OrgExample } from './pages/org'
 import { SnaplineExample } from './pages/plugins/snapline'
 import { ClipboardExample } from './pages/plugins/clipboard'
-import { KeyboardExample } from './pages/keyboard'
+import { KeyboardExample } from './pages/plugins/keyboard'
 import { DndExample } from './pages/plugins/dnd'
 import { ScrollerExample } from './pages/scroller'
 import { SelectionExample } from './pages/selection'
@@ -88,12 +88,12 @@ function App() {
       <Route path="/case/mind" element={<CaseMindExample />} />
       <Route path="/case/swimlane" element={<CaseSwimlaneExample />} />
       <Route path="/org" element={<OrgExample />} />
-      <Route path="/keyboard" element={<KeyboardExample />} />
 
       {/* Plugins Example */}
       <Route path="/plugins/dnd" element={<DndExample />} />
       <Route path="/plugins/snapline" element={<SnaplineExample />} />
       <Route path="/plugins/clipboard" element={<ClipboardExample />} />
+      <Route path="/plugins/keyboard" element={<KeyboardExample />} />
 
       <Route path="/scroller" element={<ScrollerExample />} />
       <Route path="/selection" element={<SelectionExample />} />
